都道府県別ランキングの実装

+++ supabase/migrations/20250614001740_add_get_user_mission_ranking_function.sql
@@ -6,8 +6,8 @@ returns table (
   level int,
   xp int,
   updated_at timestamp,
-  clear_count int,
-  rank int
+  clear_count bigint,
+  rank bigint
 )
 language sql
 as $$
